simple guess table
adding row to guesses table after gessing

+++ resources/views/games/mobs.blade.php
@@ -35,12 +35,25 @@
                     select: function(event, ui) {
                         console.log('Wybrano ID:', ui.item.value);
                         let mobId = ui.item.value;
-                        checkMob(mobId);
+                        check_mob(mobId);
+                        $('#mobSearch').val('');
+                        return false;
                     }
                 });
             }
         });
     });
+
+    function guesses_table(guess){
+        console.log('wywołanie dodawanie wiersza do tabeli');
+        const tbody = document.querySelector('#guesses_tab tbody');
+        
+        const row = document.createElement("tr");
+        row.innerHTML = `
+            <td>${guess.name}</td>
+        `;
+        tbody.prepend(row);
+    }
 
     function add_guess_to_ls(guess){
         mobGuesses = localStorage.getItem('mobGuesses');
@@ -54,28 +67,35 @@
                 {is_guessed: guess.is_guess_corect}
             ]
             localStorage.setItem('mobGuesses', JSON.stringify(new_data));
-            console.log("Utworzono mobGuesses", JSON.parse(mobGuesses));
+            console.log("Utworzono mobGuesses");
         } else{
+            console.log("mobGuesses istnieje");
             mobGuesses = JSON.parse(mobGuesses);
 
-            mobGuesses[1].guesses.push({name: guess.name});
-            if(guess.is_guess_corect){
-                mobGuesses[2].is_guessed = guess.is_guess_corect;
+            let alreadyguessed = mobGuesses[1].guesses.some(g => g.name === guess.name);
+            if(!alreadyguessed){
+                mobGuesses[1].guesses.push({name: guess.name});
+                if(guess.is_guess_corect){
+                    mobGuesses[2].is_guessed = guess.is_guess_corect;
+                }
+                localStorage.setItem('mobGuesses', JSON.stringify(mobGuesses));
             }
-            localStorage.setItem('mobGuesses', JSON.stringify(mobGuesses));
-            console.log("mobGuesses istnieje", mobGuesses);
+            else{
+                console.log('Już istnieje taka odpowiedź');
+            }
         }
+        
+        guesses_table(guess);
         return;
     }
 
-    function checkMob(mobId){
+    function check_mob(mobId){
         console.log('Wywołano funkcje checkMob', mobId)
         $.ajax({
             url: '/mobguesser/check_guess',
             method: 'POST',
             data: {
                 mob_id: mobId,
-                // _token: $('meta[name="csrf-token"]').attr('content')
             },
             success: function(response){
                 console.log('Odpowiedź backendu', response);
@@ -97,6 +117,19 @@
                 <div class="card-body">
                     Mobki
                     <input type="text" id="mobSearch" placeholder="Wpisz nazwę moba..." class="form-control">
+
+                    <table class="table table-bordered" id="guesses_tab">
+                        <thead class="table-active"> 
+                            <tr>
+                                <th>Name</th>
+                                <!-- <th>Health</th>
+                                <th>Height</th>
+                                <th>Nature</th> -->
+                            </tr>
+                        </thead>
+                        <tbody>
+                        </tbody>
+                    </table>
                 </div>
             </div>
         </div>
